refactor: add domain operation modules
