+++ apps/sdp-api/src/routes/issuance/handlers/mint.ts
@@ -1,11 +1,11 @@
 import type { Context } from "hono";
 import { getDb } from "@/db";
 import { parseDecimalAmount } from "@/lib/amount";
-import { resolveApiKeySigningWalletId } from "@/lib/api-key-wallet-auth";
 import { getAuth } from "@/lib/auth";
 import { AppError, notFound } from "@/lib/errors";
 import { success } from "@/lib/response";
 import { assertValidAddress } from "@/lib/solana";
+import { resolveApiKeySigningWalletId } from "@/services/api-key-scope.service";
 import { AuditService } from "@/services/audit.service";
 import { createMosaicService } from "@/services/mosaic";
 import { createOrgSigner } from "@/services/solana";
