+++ apps/sdp-api/src/routes/payments/wallets.ts
@@ -1,8 +1,8 @@
 import type { Permission } from "@sdp/types";
-import { assertApiKeyWalletAccess } from "@/lib/api-key-wallet-auth";
 import { getAuth } from "@/lib/auth";
 import { AppError } from "@/lib/errors";
 import { assertValidAddress } from "@/lib/solana";
+import { assertApiKeyWalletAccess } from "@/services/api-key-scope.service";
 import { createSigningService } from "@/services/domain/signing.service";
 import type { CustodyWallet } from "@/services/stores/custody-config.store";
 import type { AppContext } from "./context";
